# pkg/scootin/templates/openapi.yaml
openapi: 3.1.0
info:
  title: Scootin' Aboot API
  description: REST API for scooter event collection and querying.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Scooter:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [free, occupied]
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        client_id:
          type: string
          description: UUID of the client currently occupying the scooter, empty string if free
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time

  requestBodies:
    UpdateScooter:
      description: data sent from scooters
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - timestamp
            properties:
              status:
                type: string
                enum: [free, occupied]
              timestamp:
                type: string
                format: date-time
              latitude:
                type: number
                format: double
              longitude:
                type: number
                format: double
              client_id:
                type: string
                description: UUID of the client (required when occupying, must match when releasing/updating)

paths:
  /scooters/{id}:
    post:
      summary: Report a scooter event
      description: >
        Scooters send events when starting a trip, ending a trip, or updating location.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
            default: 11111111-1111-1111-1111-111111111111
      requestBody:
        $ref: "#/components/requestBodies/UpdateScooter"
      responses:
        "201":
          description: Event created successfully
        "400":
          description: Invalid input
        "401":
          description: Invalid API key

  /scooters:
    get:
      summary: Query scooters in a rectangular area
      parameters:
        - in: query
          name: latitude1
          required: true
          schema:
            type: number
            format: double
            default: -100.0
        - in: query
          name: longitude1
          required: true
          schema:
            type: number
            format: double
            default: -100.0
        - in: query
          name: latitude2
          required: true
          schema:
            type: number
            format: double
            default: 100.0
        - in: query
          name: longitude2
          required: true
          schema:
            type: number
            format: double
            default: 100.0
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [free, occupied]
      responses:
        "200":
          description: List of scooters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scooter"
        "401":
          description: Invalid API key